Fix bag

+++ src/lab1/Task2.java
@@ -2,7 +2,7 @@ package lab1;
 
 public class Task2 {
     public static void main(String[] args) {
-        System.out.println((46 + 10) * (10 / 3));
+        System.out.println((46 + 10) * ((double)10 / 3));
         System.out.println(29 * 4 * (-15));
     }
 }
