+++ src/lab1/Task3.java
@@ -3,7 +3,7 @@ package lab1;
 public class Task3 {
     public static void main(String[] args) {
         int number = 10500;
-        int result = (number / 10) / 10;
+        double result = ((number / 10) /(double) 10);
         System.out.println(result);
     }
 }
